fixed generated domain issue with seer (#121)
Seer was looking for gateways nodes only. Some clouds might not have these, or they could just not be available, seer should be able to return substrate nodes. Later, we need a smarter way to deal with it so a TODO was added.
Also, the ways generated domains was iffy with a string compiles to regex. Fixed now.